build: migrate to pnpm + TypeScript ESM/CJS/types dual build
- Convert index.js to src/vpd.ts + src/index.ts (named export calcVpd)
- Dual ESM + CJS + .d.ts output via tsc, exports map
- Switch package manager to pnpm
- BREAKING: default export -> named export, bump to v2.0.0
