+++ src/index.ts
@@ -1,0 +1,2 @@
+export type { VpdInput, VpdResult } from './vpd.js'
+export { calcVpd } from './vpd.js'
